refactor: split status command god file into focused modules (#7551)
src/commands/status.rs tripped both structural audit findings — 1558
lines (threshold 1500) and 41 top-level items (threshold 30). Convert it
to src/commands/status/mod.rs and extract three cohesive submodules:

- types.rs        — CLI args and serialized output shapes
- git_cache.rs    — StatusGitCache plus git probing/caching helpers
- context_paths.rs — registered-context detection for the default view

mod.rs keeps the orchestration entry points (run, dashboard/summary
builders) and the test module. Public surface (StatusArgs, run, output
types) is preserved via re-exports, so external callers
(command_contract descriptors) are unchanged. No behavior change; all 18
status tests pass.

Every resulting file is now under both structural thresholds.
