+++ src/commands/status/mod.rs
@@ -1,332 +1,38 @@
-use clap::Args;
+//! `homeboy status` — component/project release-state overview.
+//!
+//! This module is split into focused submodules:
+//! - [`types`] — CLI args and serialized output shapes.
+//! - [`git_cache`] — per-component git caching and probes.
+//! - [`context_paths`] — registered-context detection for the default view.
+//! - [`dashboard_table`] — human-readable dashboard rendering.
+//!
+//! The orchestration entry points (`run`, dashboard/summary builders) live
+//! here and compose those pieces.
+
 use homeboy::core::component;
 use homeboy::core::context;
-use homeboy::core::deploy::{self, ReleaseState, ReleaseStateStatus};
-use homeboy::core::git;
+use homeboy::core::deploy::ReleaseStateStatus;
 use homeboy::core::release::version;
 use homeboy::core::scope::{self, Scope};
-use serde::Serialize;
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 
 use super::CmdResult;
-use serde_json::Value;
-use std::fs;
-use std::path::{Path, PathBuf};
-use std::time::Instant;
 
+mod context_paths;
 mod dashboard_table;
+mod git_cache;
+mod types;
+
+use context_paths::unregistered_cwd_status_output;
 use dashboard_table::log_dashboard_table;
+use git_cache::{fetch_project_remote_versions, log_unreleased_merges, StatusGitCache};
 
-#[derive(Args)]
-pub struct StatusArgs {
-    /// Project ID — show version dashboard for a project's components
-    pub project: Option<String>,
-
-    /// Inspect this checkout path instead of the registered component path
-    #[arg(long, value_name = "PATH")]
-    pub path: Option<String>,
-
-    /// Show the full workspace/context report (the old init behavior)
-    #[arg(long)]
-    pub full: bool,
-
-    /// Show only components with uncommitted changes
-    #[arg(long)]
-    pub uncommitted: bool,
-
-    /// Show only components that need a release
-    #[arg(long)]
-    pub needs_release: bool,
-
-    /// Show only components ready to deploy
-    #[arg(long)]
-    pub ready: bool,
-
-    /// Show only components with docs-only changes
-    #[arg(long)]
-    pub docs_only: bool,
-
-    /// Show all components regardless of current directory context
-    #[arg(long, short = 'a')]
-    pub all: bool,
-
-    /// Show only outdated components (local != remote)
-    #[arg(long)]
-    pub outdated: bool,
-
-    /// Emit status phase progress to stderr and include phase timings in JSON
-    #[arg(long)]
-    pub timings: bool,
-
-    /// Show only components carrying merged-but-unreleased work (commits on
-    /// origin/<default-branch> that are past the latest release tag).
-    #[arg(long)]
-    pub unreleased: bool,
-}
-
-/// Per-component upstream drift info.
-#[derive(Debug, Clone, Serialize)]
-pub struct UpstreamDrift {
-    pub component_id: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub ahead: Option<u32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub behind: Option<u32>,
-    /// Latest tag on origin (e.g. "v0.8.0").
-    /// Differs from local version when the local checkout is stale.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub latest_origin_tag: Option<String>,
-}
-
-impl UpstreamDrift {
-    fn is_behind(&self) -> bool {
-        self.behind.unwrap_or(0) > 0
-    }
-}
-
-/// A component carrying code that is merged to its default branch on origin but
-/// not yet in a release tag — the "merged-not-released" state.
-///
-/// This is the complement of `ready_to_deploy`/`UpstreamDrift`: instead of
-/// asking "is the local checkout ahead of the latest tag" (which depends on a
-/// fresh local checkout and a local-vs-tag diff), it asks the higher-stakes
-/// inverse — "is there work merged on `origin/<default-branch>` that no release
-/// tag covers yet, so the code does NOT exist on prod?"
-///
-/// The count is measured against `origin/<default-branch>` (refreshed by the
-/// tag/branch fetch that already runs for upstream drift), so it is robust to a
-/// stale local HEAD. See issue #4996.
-#[derive(Debug, Clone, Serialize)]
-pub struct UnreleasedMerge {
-    pub component_id: String,
-    /// The latest release tag the unreleased work is measured against.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub latest_tag: Option<String>,
-    /// Count of commits on `origin/<default-branch>` past `latest_tag`
-    /// (merge commits excluded, matching release-state counting).
-    pub commits_since_tag: u32,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct StatusTiming {
-    pub phase: &'static str,
-    pub elapsed_ms: u128,
-}
-
-/// Clarifying note emitted alongside the git-state-only `ready_to_deploy` list.
-///
-/// `ready_to_deploy` reflects local git/workspace state ("has a clean release
-/// tag that *could* be deployed"), NOT whether the deploy target is actually
-/// behind. Acting on it blindly re-deploys components that may already be live.
-/// For a target-accurate diff (installed version vs latest release tag) run
-/// `homeboy status <project>`, which reports `current` / `outdated` per
-/// component. See issue #4588.
-const READY_TO_DEPLOY_NOTE: &str = "ready_to_deploy is git-state-only (components with a clean release tag that *could* be deployed); it does NOT mean the deploy target is behind. Run `homeboy status <project>` for a target-accurate diff (installed version vs latest release tag).";
-
-/// Clarifying note emitted alongside `unreleased_merges`.
-///
-/// `unreleased_merges` flags components whose `origin/<default-branch>` carries
-/// commits past the latest release tag — i.e. work that is merged but not in any
-/// release, so the code does NOT exist on prod. This is the merged→released axis
-/// of the merged→released→deployed chain; for the released→deployed axis
-/// (installed version vs latest tag) run `homeboy status <project>`. See #4996.
-const UNRELEASED_MERGES_NOTE: &str = "unreleased_merges flags components with commits merged to origin/<default-branch> that are past the latest release tag (merged but NOT released — the code is not on prod yet). A merged PR here is NOT 'shipped'. Cut a release, then run `homeboy status <project>` to confirm installed-vs-tag (released-but-not-deployed).";
-
-#[derive(Debug, Serialize)]
-pub struct StatusOutput {
-    pub command: &'static str,
-    pub total: usize,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub uncommitted: Vec<String>,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub needs_release: Vec<String>,
-    /// Components in a clean release state (no uncommitted changes, no commits
-    /// since the last version tag).
-    ///
-    /// IMPORTANT: this is git/workspace state only. It means "has a release tag
-    /// that *could* be deployed", NOT "the deployed target is behind the latest
-    /// release". For a target-accurate deploy diff, use `homeboy status
-    /// <project>` (compares installed-on-target versions against release tags).
-    /// See `ready_to_deploy_note` and issue #4588.
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub ready_to_deploy: Vec<String>,
-    /// Clarifying note, emitted only when `ready_to_deploy` is non-empty, so
-    /// operators don't mistake the git-state list for a real deploy backlog.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub ready_to_deploy_note: Option<&'static str>,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub docs_only: Vec<String>,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub behind_upstream: Vec<String>,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub upstream_drift: Vec<UpstreamDrift>,
-    /// Components carrying code merged to `origin/<default-branch>` but not yet
-    /// covered by a release tag ("merged-not-released"). Closes the false
-    /// "shipped" read where a merged PR is mistaken for live code. Additive and
-    /// independent of `ready_to_deploy` (issue #4996).
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub unreleased_merges: Vec<UnreleasedMerge>,
-    /// Clarifying note, emitted only when `unreleased_merges` is non-empty.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub unreleased_merges_note: Option<&'static str>,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub timings: Vec<StatusTiming>,
-    pub clean: usize,
-}
-
-#[derive(Debug, Serialize)]
-pub struct UnregisteredContextStatusOutput {
-    pub command: &'static str,
-    pub status: &'static str,
-    pub cwd: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub git_root: Option<String>,
-    pub suggestion: String,
-    pub action: &'static str,
-}
-
-/// A single row in the project status dashboard.
-#[derive(Debug, Serialize)]
-pub struct ProjectStatusRow {
-    pub component_id: String,
-    pub local_version: Option<String>,
-    pub remote_version: Option<String>,
-    /// Actionable evidence when the bounded remote version probe failed.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub remote_version_diagnostic: Option<String>,
-    /// Latest tag on origin. When this differs from local_version, the local
-    /// checkout is stale and needs a pull.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub origin_version: Option<String>,
-    pub unreleased_commits: u32,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub ahead_upstream: Option<u32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub behind_upstream: Option<u32>,
-    pub status: ProjectComponentDashboardStatus,
-}
-
-/// Status indicator for the project dashboard.
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ProjectComponentDashboardStatus {
-    /// Local and remote versions match, no unreleased commits
-    Current,
-    /// Local and remote versions match, but a newer origin tag exists
-    PinnedCurrent,
-    /// Local version differs from remote (needs deploy)
-    Outdated,
-    /// Releasable code commits since the current version baseline
-    NeedsRelease,
-    /// Only docs changes since last tag
-    DocsOnly,
-    /// Uncommitted changes in working directory
-    Uncommitted,
-    /// Local branch is behind upstream (needs pull)
-    BehindUpstream,
-    /// Absorbed into another component; not independently deployable. Tracked
-    /// for visibility but excluded from deploy/outdated obligations.
-    Bundled,
-    /// Sunset; no longer a deploy target. Excluded from outdated obligations.
-    Retired,
-    /// Cannot determine status
-    Unknown,
-    /// A bounded remote probe failed; local and git data remain available.
-    Degraded,
-}
-
-/// Output for the project status dashboard.
-#[derive(Debug, Serialize)]
-pub struct ProjectDashboardOutput {
-    pub command: &'static str,
-    pub project_id: String,
-    pub total: usize,
-    pub components: Vec<ProjectStatusRow>,
-    pub summary: ProjectDashboardSummary,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub timings: Vec<StatusTiming>,
-}
-
-struct StatusTimer {
-    enabled: bool,
-    phase_started: Instant,
-    timings: Vec<StatusTiming>,
-}
-
-impl StatusTimer {
-    fn new(enabled: bool) -> Self {
-        Self {
-            enabled,
-            phase_started: Instant::now(),
-            timings: Vec::new(),
-        }
-    }
-
-    fn begin(&mut self, phase: &'static str) {
-        if self.enabled {
-            eprintln!("[status] {phase}...");
-        }
-        self.phase_started = Instant::now();
-    }
-
-    fn finish(&mut self, phase: &'static str) {
-        if !self.enabled {
-            return;
-        }
-
-        let elapsed_ms = self.phase_started.elapsed().as_millis();
-        eprintln!("[status] {phase} completed in {elapsed_ms}ms");
-        self.timings.push(StatusTiming { phase, elapsed_ms });
-    }
-
-    fn into_timings(self) -> Vec<StatusTiming> {
-        self.timings
-    }
-}
-
-/// Summary counts for the project dashboard.
-#[derive(Debug, Serialize)]
-pub struct ProjectDashboardSummary {
-    pub current: usize,
-    pub pinned_current: usize,
-    pub outdated: usize,
-    pub needs_release: usize,
-    pub docs_only: usize,
-    pub uncommitted: usize,
-    pub behind_upstream: usize,
-    #[serde(default, skip_serializing_if = "is_zero")]
-    pub bundled: usize,
-    #[serde(default, skip_serializing_if = "is_zero")]
-    pub retired: usize,
-    pub unknown: usize,
-    #[serde(default, skip_serializing_if = "is_zero")]
-    pub degraded: usize,
-}
-
-fn is_zero(value: &usize) -> bool {
-    *value == 0
-}
-
-pub enum StatusResult {
-    Summary(StatusOutput),
-    UnregisteredContext(UnregisteredContextStatusOutput),
-    Full(homeboy::core::context::report::ContextReport),
-    Dashboard(ProjectDashboardOutput),
-}
-
-impl serde::Serialize for StatusResult {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        match self {
-            StatusResult::Summary(output) => output.serialize(serializer),
-            StatusResult::UnregisteredContext(output) => output.serialize(serializer),
-            StatusResult::Full(output) => output.serialize(serializer),
-            StatusResult::Dashboard(output) => output.serialize(serializer),
-        }
-    }
-}
+pub use types::{
+    ProjectComponentDashboardStatus, ProjectDashboardOutput, ProjectDashboardSummary,
+    ProjectStatusRow, StatusArgs, StatusOutput, StatusResult, StatusTiming,
+    UnregisteredContextStatusOutput, UnreleasedMerge, UpstreamDrift,
+};
+use types::{StatusTimer, READY_TO_DEPLOY_NOTE, UNRELEASED_MERGES_NOTE};
 
 pub fn run(args: StatusArgs, _global: &super::GlobalArgs) -> CmdResult<StatusResult> {
     if args.path.is_some() {
@@ -396,118 +102,6 @@ pub fn run(args: StatusArgs, _global: &super::GlobalArgs) -> CmdResult<StatusRes
     };
 
     summarize_components(components, &args, timer)
-}
-
-fn unregistered_cwd_status_output() -> Option<UnregisteredContextStatusOutput> {
-    let cwd = std::env::current_dir().ok()?;
-    let git_root = git::get_git_root(&cwd.to_string_lossy())
-        .ok()
-        .map(PathBuf::from);
-    let candidates = [Some(cwd.as_path()), git_root.as_deref()];
-
-    if candidates
-        .into_iter()
-        .flatten()
-        .any(path_is_registered_context)
-    {
-        return None;
-    }
-
-    let git_root_string = git_root
-        .as_ref()
-        .map(|path| path.to_string_lossy().to_string());
-    let suggestion = if let Some(ref git_root) = git_root_string {
-        format!(
-            "Repo not attached. Prefer: `homeboy project components attach-path <project-id> {}`",
-            git_root
-        )
-    } else {
-        "Repo not attached. Prefer: `homeboy project components attach-path <project-id> <path>`"
-            .to_string()
-    };
-
-    Some(UnregisteredContextStatusOutput {
-        command: "status",
-        status: "unregistered_context",
-        cwd: cwd.to_string_lossy().to_string(),
-        git_root: git_root_string,
-        suggestion,
-        action: "Run `homeboy status --all` to inspect every configured component, or attach this checkout to a project/component first.",
-    })
-}
-
-fn path_is_registered_context(path: &Path) -> bool {
-    registered_local_paths().into_iter().any(|registered| {
-        path_is_at_or_inside(&registered, path) || path_is_at_or_inside(path, &registered)
-    })
-}
-
-fn registered_local_paths() -> Vec<PathBuf> {
-    let Ok(home) = std::env::var("HOME") else {
-        return Vec::new();
-    };
-    let config_root = PathBuf::from(home).join(".config").join("homeboy");
-    [config_root.join("components"), config_root.join("projects")]
-        .into_iter()
-        .flat_map(json_files_under)
-        .filter_map(|path| fs::read_to_string(path).ok())
-        .filter_map(|raw| serde_json::from_str::<Value>(&raw).ok())
-        .flat_map(|value| {
-            let mut paths = Vec::new();
-            collect_local_paths(&value, &mut paths);
-            paths
-        })
-        .collect()
-}
-
-fn json_files_under(root: PathBuf) -> Vec<PathBuf> {
-    let mut files = Vec::new();
-    collect_json_files(&root, &mut files);
-    files
-}
-
-fn collect_json_files(path: &Path, files: &mut Vec<PathBuf>) {
-    if path.is_file() {
-        if path.extension().and_then(|ext| ext.to_str()) == Some("json") {
-            files.push(path.to_path_buf());
-        }
-        return;
-    }
-
-    let Ok(entries) = fs::read_dir(path) else {
-        return;
-    };
-    for entry in entries.flatten() {
-        collect_json_files(&entry.path(), files);
-    }
-}
-
-fn collect_local_paths(value: &Value, paths: &mut Vec<PathBuf>) {
-    match value {
-        Value::Object(map) => {
-            for (key, value) in map {
-                if matches!(key.as_str(), "local_path" | "localPath") {
-                    if let Some(path) = value.as_str().filter(|path| !path.trim().is_empty()) {
-                        paths.push(homeboy::core::expand_tilde_path(path));
-                    }
-                }
-                collect_local_paths(value, paths);
-            }
-        }
-        Value::Array(items) => {
-            for item in items {
-                collect_local_paths(item, paths);
-            }
-        }
-        _ => {}
-    }
-}
-
-fn path_is_at_or_inside(parent: &Path, path: &Path) -> bool {
-    match (parent.canonicalize().ok(), path.canonicalize().ok()) {
-        (Some(parent), Some(path)) => path == parent || path.starts_with(parent),
-        _ => false,
-    }
 }
 
 fn summarize_components(
@@ -895,246 +489,16 @@ fn origin_tag_is_newer_than_local(origin_tag: Option<&str>, local: &str) -> bool
         .is_some_and(|(origin, local)| origin > local)
 }
 
-#[derive(Default)]
-struct StatusGitCache {
-    upstream_drift: HashMap<String, Option<UpstreamDrift>>,
-    fetched_tags: HashSet<String>,
-    release_states: HashMap<String, Option<ReleaseState>>,
-    baselines: HashMap<String, Option<git::BaselineInfo>>,
-    origin_branches: HashMap<String, Option<String>>,
-}
-
-impl StatusGitCache {
-    fn fetch_origin_tags_for(&mut self, path: &str) {
-        let cache_key = upstream_drift_cache_key(path);
-        if self.fetched_tags.insert(cache_key) {
-            fetch_origin_tags(path);
-        }
-    }
-
-    fn fetch_upstream_drift_for(
-        &mut self,
-        component: &component::Component,
-    ) -> Option<UpstreamDrift> {
-        let path = &component.local_path;
-        let cache_key = component_cache_key(component);
-        if !self.upstream_drift.contains_key(&cache_key) {
-            self.fetch_origin_tags_for(path);
-            self.upstream_drift
-                .insert(cache_key.clone(), get_upstream_drift(component));
-        }
-
-        let drift = self.upstream_drift.get(&cache_key)?;
-
-        drift.as_ref().map(|cached| {
-            let mut drift = cached.clone();
-            drift.component_id = component.id.clone();
-            drift
-        })
-    }
-
-    fn release_state_for(&mut self, component: &component::Component) -> Option<&ReleaseState> {
-        let cache_key = component_cache_key(component);
-        if !self.release_states.contains_key(&cache_key) {
-            let state = self.baseline_for(component).and_then(|baseline| {
-                deploy::calculate_release_state_from_baseline(component, baseline)
-            });
-            self.release_states.insert(cache_key.clone(), state);
-        }
-
-        self.release_states.get(&cache_key).and_then(Option::as_ref)
-    }
-
-    fn baseline_for(&mut self, component: &component::Component) -> Option<&git::BaselineInfo> {
-        let cache_key = component_cache_key(component);
-        if !self.baselines.contains_key(&cache_key) {
-            self.fetch_origin_tags_for(&component.local_path);
-            let current_version = version::read_component_version(component)
-                .ok()
-                .map(|info| info.version);
-            let tag_prefix = homeboy::core::release::component_tag_prefix(component)
-                .ok()
-                .flatten();
-            let baseline = git::detect_baseline_with_version_and_tag_prefix_from_fetched_tags(
-                &component.local_path,
-                current_version.as_deref(),
-                tag_prefix.as_deref(),
-            )
-            .ok();
-            self.baselines.insert(cache_key.clone(), baseline);
-        }
-
-        self.baselines.get(&cache_key).and_then(Option::as_ref)
-    }
-
-    fn default_origin_branch_for(&mut self, path: &str) -> Option<&str> {
-        let cache_key = upstream_drift_cache_key(path);
-        if !self.origin_branches.contains_key(&cache_key) {
-            self.origin_branches
-                .insert(cache_key.clone(), default_origin_branch(path));
-        }
-
-        self.origin_branches
-            .get(&cache_key)
-            .and_then(Option::as_deref)
-    }
-
-    fn detect_unreleased_merges_for(
-        &mut self,
-        comp: &component::Component,
-    ) -> Option<UnreleasedMerge> {
-        let path = &comp.local_path;
-
-        let origin_branch = self.default_origin_branch_for(path)?.to_string();
-        let baseline = self.baseline_for(comp)?;
-        let baseline_ref = baseline.reference.as_deref()?;
-
-        let range = format!("{}..{}", baseline_ref, origin_branch);
-        let count_output = homeboy::core::engine::command::run_in_optional(
-            path,
-            "git",
-            &["rev-list", "--count", "--no-merges", &range],
-        )?;
-
-        let commits_since_tag: u32 = count_output.trim().parse().ok()?;
-        if commits_since_tag == 0 {
-            return None;
-        }
-
-        Some(UnreleasedMerge {
-            component_id: comp.id.clone(),
-            latest_tag: baseline.latest_tag.clone(),
-            commits_since_tag,
-        })
-    }
-}
-
-fn upstream_drift_cache_key(path: &str) -> String {
-    git::get_git_root(path).unwrap_or_else(|_| path.to_string())
-}
-
-fn component_cache_key(component: &component::Component) -> String {
-    format!("{}\0{}", component.id, component.local_path)
-}
-
-fn fetch_origin_tags(path: &str) {
-    // Best-effort fetch — silently proceeds if no remote or network issue.
-    let _ = homeboy::core::engine::command::run_in_optional(
-        path,
-        "git",
-        &["fetch", "--tags", "--quiet"],
-    );
-}
-
-fn get_upstream_drift(component: &component::Component) -> Option<UpstreamDrift> {
-    let path = &component.local_path;
-    let snapshot = git::get_repo_snapshot(path).ok()?;
-
-    // After fetching tags, find the latest tag across ALL refs (not just HEAD).
-    // `git describe --tags --abbrev=0` only returns tags reachable from HEAD,
-    // which misses newer tags when the local checkout is behind.
-    let tag_prefix = homeboy::core::release::component_tag_prefix(component)
-        .ok()
-        .flatten();
-    let latest_origin_tag = git::get_latest_tag_any_with_prefix(path, tag_prefix.as_deref())
-        .ok()
-        .flatten();
-
-    Some(UpstreamDrift {
-        component_id: String::new(), // caller sets component_id after
-        ahead: snapshot.ahead,
-        behind: snapshot.behind,
-        latest_origin_tag,
-    })
-}
-
-/// Log merged-but-unreleased components to stderr for human-readable output.
-///
-/// Mirrors the dashboard table's terminal-only behavior so JSON consumers are
-/// unaffected. Keeps the merged-not-released signal visible in `homeboy status`
-/// without a project argument (issue #4996).
-fn log_unreleased_merges(merges: &[UnreleasedMerge]) {
-    if merges.is_empty() || !std::io::IsTerminal::is_terminal(&std::io::stderr()) {
-        return;
-    }
-
-    eprintln!(
-        "⚠️  {} component(s) carry merged-but-unreleased work (merged to origin, NOT in any release — code is not on prod yet):",
-        merges.len()
-    );
-    for merge in merges {
-        let tag = merge.latest_tag.as_deref().unwrap_or("(no tag)");
-        eprintln!(
-            "    {} — {} commit(s) past {}",
-            merge.component_id, merge.commits_since_tag, tag
-        );
-    }
-    eprintln!("    Cut a release, then `homeboy status <project>` to confirm installed-vs-tag.");
-}
-
-/// Resolve the default origin branch ref for a checkout.
-///
-/// Precedence matches the deploy planner: `origin/HEAD` symbolic ref first, then
-/// the conventional `origin/main` / `origin/trunk` / `origin/master` fallbacks.
-fn default_origin_branch(path: &str) -> Option<String> {
-    if let Some(symbolic) = homeboy::core::engine::command::run_in_optional(
-        path,
-        "git",
-        &[
-            "symbolic-ref",
-            "--quiet",
-            "--short",
-            "refs/remotes/origin/HEAD",
-        ],
-    ) {
-        let symbolic = symbolic.trim();
-        if !symbolic.is_empty() {
-            return Some(symbolic.to_string());
-        }
-    }
-
-    ["origin/main", "origin/trunk", "origin/master"]
-        .iter()
-        .find(|branch| {
-            homeboy::core::engine::command::run_in_optional(
-                path,
-                "git",
-                &["rev-parse", "--verify", "--quiet", branch],
-            )
-            .is_some()
-        })
-        .map(|branch| (*branch).to_string())
-}
-
-/// Fetch remote (deployed) versions for all components in a project.
-///
-/// Uses deploy check mode internally, which handles SSH resolution.
-/// Returns empty map on failure (e.g., no server configured, SSH unavailable).
-fn fetch_project_remote_versions(
-    project_id: &str,
-    components: &[component::Component],
-) -> deploy::RemoteVersionProbeResult {
-    match deploy::fetch_project_remote_versions(project_id, components) {
-        Ok(result) => result,
-        Err(_) => {
-            homeboy::log_status!(
-                "status",
-                "Warning: could not fetch remote versions for project '{}' — showing local data only",
-                project_id
-            );
-            deploy::RemoteVersionProbeResult::default()
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
+    use super::git_cache::{component_cache_key, default_origin_branch, upstream_drift_cache_key};
     use super::*;
     use crate::cli_surface::{Cli, Commands};
     use crate::commands::GlobalArgs;
     use clap::Parser;
     use std::env;
     use std::fs;
+    use std::path::PathBuf;
     use std::process::Command;
     use std::sync::{Mutex, OnceLock};
     use std::time::Instant;
